requirements.txt
asgiref==3.3.1
astroid==2.5.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.7
django-allauth==0.44.0
django-ckeditor==6.0.0
# Editable install with no version control (django-import-export==2.5.0)
-e /Library/Frameworks/Python.framework/Versions/3.9/lib/python3.9/site-packages
django-js-asset==1.2.2
et-xmlfile==1.0.1
gunicorn==20.1.0
idna==2.10
isort==5.7.0
lazy-object-proxy==1.5.2
MarkupPy==1.14
mccabe==0.6.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.7
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
pylint==2.7.2
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
static3==0.7.0
tablib==3.0.0
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
